+++ src/index.js
@@ -26,6 +26,8 @@ import {
   wpListPagesToolDefinition,
   wpListCategoriesToolDefinition,
   wpListTagsToolDefinition,
+  wpCreateCategoryToolDefinition,
+  wpCreateTagToolDefinition,
   wpListMenusToolDefinition,
   wpListMenuItemsToolDefinition,
   wpCreatePostToolDefinition,
@@ -37,6 +39,8 @@ import {
   handleWpListPages,
   handleWpListCategories,
   handleWpListTags,
+  handleWpCreateCategory,
+  handleWpCreateTag,
   handleWpListMenus,
   handleWpListMenuItems,
   handleWpCreatePost,
@@ -89,6 +93,8 @@ const TOOLS = [
   wpListPagesToolDefinition,
   wpListCategoriesToolDefinition,
   wpListTagsToolDefinition,
+  wpCreateCategoryToolDefinition,
+  wpCreateTagToolDefinition,
   wpListMenusToolDefinition,
   wpListMenuItemsToolDefinition,
   wpCreatePostToolDefinition,
@@ -135,6 +141,8 @@ server.setRequestHandler(CallToolRequestSchema, async (request) => {
       case "wordpress_list_pages":        return await handleWpListPages(args);
       case "wordpress_list_categories":   return await handleWpListCategories(args);
       case "wordpress_list_tags":         return await handleWpListTags(args);
+      case "wordpress_create_category":   return await handleWpCreateCategory(args);
+      case "wordpress_create_tag":        return await handleWpCreateTag(args);
       case "wordpress_list_menus":        return await handleWpListMenus(args);
       case "wordpress_list_menu_items":   return await handleWpListMenuItems(args);
       case "wordpress_create_post":       return await handleWpCreatePost(args);